requirements.txt
pytest
pylint
coverage
pylint-gitlab
pytest-cov
decorators
mock
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
googlemaps
python-telegram-bot
pytest-mock
flask
requests
pyngrok